Rails 5
Necessary changes for Rails 5.

There is a new table that rails adds to the database:
`ar_internal_metadata`
Not unlike `schema_migrations` we need to add this to the ignore list as
it contains no id column.

The format of the payload passed to `add_new_locks_held` has changed.
The old format:
```
 [[#<ActiveRecord::ConnectionAdapters::PostgreSQLColumn:0x007fe3ef8e1e78
   @array=false,
   @cast_type=#<ActiveRecord::ConnectionAdapters::PostgreSQL::OID::Uuid:0x007fe3eccaa120
@limit=nil, @precision=nil, @scale=nil>,
   @default=nil,
   @default_function="uuid_generate_v4()",
   @name="id",
   @null=false,
   @sql_type="uuid">,
  "a97839f3-daa0-4d9c-b9fe-47a7e403fe50"]]
```

The new format:
```
[#<ActiveRecord::Relation::QueryAttribute:0x007fa4e144fcb8
  @name="id",
  @original_attribute=nil,
  @type=#<ActiveRecord::ConnectionAdapters::PostgreSQL::OID::Uuid:0x007fa4dbbbe720
@limit=nil, @precision=nil, @scale=nil>,
  @value="a97839f3-daa0-4d9c-b9fe-47a7e403fe50",
  @value_before_type_cast="a97839f3-daa0-4d9c-b9fe-47a7e403fe50",
  @value_for_database="a97839f3-daa0-4d9c-b9fe-47a7e403fe50">]
```